Tests: no need to pass --ignore=:iconv: to all tests
Now that iconv is not used unless needed, the tests, none of which require
it, do not need to turn it off; and indeed, we want to keep it on, to make
sure it does not interfere with the results!

The one exception is t21_names.py, which lists the known character sets; for
its result to be consistent with the standard result, we must disable iconv
when we run it.