Fix output of request diagnostics with --verbose
Since commit ad92ff8 (Only use iconv if we need it), we have called
scan_request() twice in some circumstances. When --verbose is used, this
outputs diagnostics about the request. So in the best case the diagnostics
are output twice, and in the worst case they are different, and the user
doesn’t know which is being used!

Hence, don’t set request_option.verbose_flag, and instead, scan the request
one final time, setting request->verbose_flag if --verbose was used. Since
we are thereby re-scanning the already-scanned request with the correct
options, this is bound to succeed, but this time will print any diagnostics.

This is inefficient, but avoids changing the public API recode_scan_request,
and only a one-off cost per invocation of recode.