Remove pipe mode of operation
This was extra complexity, and created a problem; see
https://github.com/rrthomas/recode/issues/67

In short, librecode forking could cause output from the calling process to
be duplicated, as stdio buffers would be flushed when a child process ended,
and then flushed again later in the (duplicate) copy in the caller. This is
hard to get around nicely.

Since there’s little evidence that the multi-process operation was bringing
significant benefits, and since a duplicate bug that had not been fixed on
the non-pipe code path was recently spotted as well, remove the pipe code
path.